+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1] - 2019-11-30
+
 ### Fixed
 
 - Fix cursor position after Extract Variable that creates a shorthand on an object property that ends with a comma, so rename works
@@ -593,7 +595,8 @@ function doSomething(someData) {
 
 </details>
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/1.2.0...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/1.2.1...HEAD
+[1.2.1]: https://github.com/nicoespeon/abracadabra/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/nicoespeon/abracadabra/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/nicoespeon/abracadabra/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/nicoespeon/abracadabra/compare/1.0.0...1.0.1